tests: tap: remove semicolons in pass / fail definitions
I wanted to write something like this in a test:

    if (cond)
      fail("message 1");
    else
      fail("message 2");

And was met with

      CC       test_action.o
    /home/simark/src/lttng-tools/tests/unit/test_action.c: In function ‘main’:
    /home/simark/src/lttng-tools/tests/unit/test_action.c:544:9: error: ‘else’ without a previous ‘if’
      544 |         else
          |         ^~~~

I then remembered that it was in our coding style to use braces:

    if (cond) {
      fail("message 1");
    } else {
      fail("message 2");
    }

... which avoids the error.  But still, the former form should work.
Fix this by removing the semi-colons in the pass and fail definitions, I
don't think they belong there.  Doing so finds a spot in ini_config.c
where a semi-colon is missing, add it.

Change-Id: I6ff09d496a0b12f34baa6f993cffc69eef611df0
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>