common/macros: don't define min and max macros in C++
Later in the series, the min macro conflicted with a C++ header file.
Since the min and max macros are not useful in C++ (std::min/std::max
are preferred, don't define them when building a C++ file.  Don't define
min_t and max_t either, they too can be replaced with std::min and
std::max, which are templated / type-safe.

Change-Id: I3d56d325f6508c32baba674c335c3f4ab0ecc582
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>