lib: compile liblttng-ctl as C++
Same as the previous commits, but compile the liblttng-ctl library as
C++ code (while still offering a C interface).

Some exported global variables (for example in deprecated-symbols.cpp)
have to be made non-const, otherwise we get:

      CXX      deprecated-symbols.lo
    /home/simark/src/lttng-tools/src/lib/lttng-ctl/deprecated-symbols.cpp:21:33: error: ‘visibility’ attribute ignored [-Werror=attributes]
       21 | LTTNG_EXPORT const char * const config_element_pid_tracker;
          |                                 ^~~~~~~~~~~~~~~~~~~~~~~~~~

I think this is related to the fact that const global variables
automatically have internal linkage in C++.

Despite using -export-symbols in src/lib/lttng-ctl/Makefile.am, some new
ELF symbols become exposed.  It could be related to this, in the ld man
page:

           --retain-symbols-file does not discard undefined symbols, or
           symbols needed for relocations.

One new symbol I see, for example, is `_Z16connect_sessiondv`.  Looking
at liblttng-ctl.so, I indeed see a relocation for this symbol:

    000000000010b778  0000053e00000007 R_X86_64_JUMP_SLOT 00000000000314a2 _Z16connect_sessiondv + 0

And that would explain why the linker keeps that symbol visible.  This
is related to the entry for this function in the procedure linkage
table. I'm not entirely sure why these functions didn't generate a PLT
entry in C, but do in C++.

To avoid these new symbols, build everything with -fvisibility=hidden
and -fvisibility-inlines-hidden, and tag functions we really want to
export with LTTNG_EXPORT, a new macro defined to
__attribute__((visibility("default"))).  This macro is publicly visible,
because it has to be used in distributed header files (although it's of
no use for users of liblttng-ctl).

Change-Id: Ie51bf0a2edfb87e5f46f9c39eed5309d9f8c41d6
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>